Update mine_items.dm
Добавлен do_after + перезарядка фороном и сумки шахтера

+++ code/modules/mining/mine_items.dm
@@ -714,6 +714,8 @@ var/global/mining_shuttle_location = 0 // 0 = station 13, 1 = mining station
 			to_chat(user,"<span class='notice'>[src] is already fully charged.</span>")
 			return
 		var/obj/item/stack/sheet/S = A
+		if(!do_after(user, 10, TRUE, S, can_move = TRUE))
+			return
 		S.use(1)
 		power_supply.give(1000)
 		to_chat(user, "<span class='notice'>You insert [A] in [src], recharging it.</span>")
@@ -721,9 +723,25 @@ var/global/mining_shuttle_location = 0 // 0 = station 13, 1 = mining station
 		if(power_supply.charge >= power_supply.maxcharge)
 			to_chat(user,"<span class='notice'>[src] is already fully charged.</span>")
 			return
+		if(!do_after(user, 10, TRUE, A, can_move = TRUE))
+			return
 		qdel(A)
 		power_supply.give(500)
 		to_chat(user, "<span class='notice'>You insert [A] in [src], recharging it.</span>")
+	else if(istype(A, /obj/item/weapon/storage/bag/ore))
+		if(power_supply.charge >= power_supply.maxcharge)
+			to_chat(user,"<span class='notice'>[src] is already fully charged.</span>")
+			return
+		if(!do_after(user, 10, TRUE, A, can_move = TRUE))
+			return
+		var/obj/item/weapon/storage/bag/ore/O = A
+		for(var/obj/item/weapon/ore/phoron/P in O.contents)
+			if(power_supply.charge >= power_supply.maxcharge)
+				return
+			O.remove_from_storage(P)
+			qdel(P)
+			power_supply.give(500)
+			to_chat(user, "<span class='notice'>You insert [P] in [src], recharging it.</span>")
 	else
 		return ..()
